move FillLogResult to types file to allow usage in client
