+++ client/ts/src/fillFeed.ts
@@ -8,6 +8,7 @@ import { genAccDiscriminator } from './utils/discriminator';
 import * as promClient from 'prom-client';
 import express from 'express';
 import promBundle from 'express-prom-bundle';
+import { FillLogResult } from './types';
 
 // For live monitoring of the fill feed. For a more complete look at fill
 // history stats, need to index all trades.
@@ -170,33 +171,6 @@ export async function runFillFeed() {
 
 const fillDiscriminant = genAccDiscriminator('manifest::logs::FillLog');
 
-/**
- * FillLogResult is the message sent to subscribers of the FillFeed
- */
-export type FillLogResult = {
-  /** Public key for the market as base58. */
-  market: string;
-  /** Public key for the maker as base58. */
-  maker: string;
-  /** Public key for the taker as base58. */
-  taker: string;
-  /** Number of base atoms traded. */
-  baseAtoms: string;
-  /** Number of quote atoms traded. */
-  quoteAtoms: string;
-  /** Price as float. Quote atoms per base atom. */
-  price: number;
-  /** Boolean to indicate which side the trade was. */
-  takerIsBuy: boolean;
-  /** Boolean to indicate whether the maker side is global. */
-  isMakerGlobal: boolean;
-  /** Sequential number for every order placed / matched wraps around at u64::MAX */
-  makerSequenceNumber: string;
-  /** Sequential number for every order placed / matched wraps around at u64::MAX */
-  takerSequenceNumber: string;
-  /** Slot number of the fill. */
-  slot: number;
-};
 function toFillLogResult(fillLog: FillLog, slot: number): FillLogResult {
   return {
     market: fillLog.market.toBase58(),
