+++ client/ts/src/index.ts
@@ -1,5 +1,6 @@
 export * from './client';
 export * from './market';
+export * from './types';
 // Do not export all of manifest because names collide with wrapper. Force users
 // to use the client.
 export * from './manifest/errors';
